Fix bug where default value will not be read.
If the key does not exist in the idb, the default value would previously be overwritten by an undefined value. This is, in my opinion, not desirable.
Apart form that, the useEffect API does not support an async function directly and would throw an error.
This fixes these two bugs.